Minor Debugs
1.) Macro Added to fix problems with old versions of g++
2.) Removed ToHex.cpp and transferred its function to ReadAndGetContent
